Got smooth movement calculations from input working with a following camera (under the hood, no visuals).

+++ src/game/ecs/components/target_position.rs
@@ -1,9 +1,9 @@
 use specs::{Component, VecStorage};
 use specs_derive::Component;
 
-#[derive(Default, Component)]
+#[derive(Default, Component, Debug)]
 #[storage(VecStorage)]
 pub struct TargetPosition {
-    pub x: f32,
-    pub y: f32,
+    pub x: u32,
+    pub y: u32,
 }
